Featured Datasets

Access featured datasets and associated metadata. These datasets required substantial preparation to enhance their utility for research and analysis.

Usage

featured_data(dataset_id, ...)

featured_metadata()

featured_ids()

Arguments

dataset_id

ID for a particular featured dataset. Run featured_ids() for the list of available options.

...

Character vectors for filtering the data on specific columns, with one vector for each column. See examples below for the syntax for different types of columns. The following list gives instructions for columns that might be available for your dataset. If they are not available you will get a message telling you what columns are in the dataset that you are asking for.

location : Location descriptor as reported in the original source.

iso_3166 : ISO-3166 codes for locations corresponding to countries.

iso_3166_2 : ISO-3166-2 codes for locations corresponding to sub-national jurisdictions.

date : Date of time point at which a value is reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.

period_start_date : The first date within the time interval for which a count was reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.

period_end_date : The last date within the time interval for which a count was reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.

period_mid_date : Date in the (rounded down) middle of a time interval for which a count was reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.

historical_disease_family : Disease family as reported by the original source.

historical_disease : Disease name as reported by the original source.

disease : Disease name harmonized over data sources.

nesting_disease : Disease name that encapsulates this disease in a hierarchy of diseases.

basal_disease : Base disease name that encapsulates this disease in a hierarchy of diseases.

icd_9 : Disease codes for version 9 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.

icd_7 : Disease codes for version 7 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.

historical_disease_subclass : Disease subclass as reported by the original source – cf. Disease

icd_9_subclass : Disease codes associated with a subclass identified by the original source. Codes are for version 9 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.

icd_7_subclass : Disease codes associated with a subclass identified by the original source. Codes are for version 7 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.

lower_age : Lower age bound in years of all individuals associated with a particular record. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

upper_age : Upper age bound in years of all individuals associated with a particular record. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

sex : Sex of all individuals associated with a particular record as reported by the original source.

cases_this_period : Number of cases reported for this period, at the end of this period – as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

cases_prev_period : Number of cases reported for the previous period, at the end of this current period – as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

cases_cum_report_year : Total number of cases up until this current period, in the current year – as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

cases_cum_prev_year : Total number of cases up until this current period, in the previous year – as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with.

cases_median_prev_5_years : Median number of weekly cases in this current period, with the median taken over the previous five years – as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

cases_cum_median_prev_5_years : Median number of cumulative cases by this current period, with the median taken over the previous five years – as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

population : Estimated total number of individuals associated with a particular record. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.

cause : Cause of mortality as reported by the original source.

location_type : Location type of recorded data (continents, countries, canadian-provinces, canadian-cities, ontario-cities, uk-cities, london-parishes) for a given record.

time_scale : Time scale of recorded data (wk, 2wk, mo, qr, yr) for a given record. This field is a qualitative and approximate measure of the time scale over which data were counted. For example, monthly (mo) data might include four-weekly or even five-weekly data. See columns like days_this_period for a quantitative description of the time scale if it is available, or produce your own description using period_start_date and period_end_date.

Functions

featured_data(): Return a data frame of a featured dataset, possibly filtered.

featured_metadata(): Return a list of lists, each of which represents the metadata for a featured dataset.

featured_ids(): Return the IDs of all the featured datasets in the repository.

Examples

options(iidda_api_msgs = FALSE)
featured_ids()
#> [1] "canmod-cdi-harmonized"   "canmod-cdi-normalized"  
#> [3] "canmod-cdi-unharmonized" "canmod-mort-harmonized" 
#> [5] "canmod-mort-normalized"  "canmod-pop-normalized"  
atlantic_polio_1950s = featured_data("canmod-cdi-normalized"
  , iso_3166_2 = c("CA-NL", "CA-NS", "CA-PE", "CA-NB")
  , basal_disease = "poliomyelitis"
  , period_end_date = "1950-01-01..1959-12-31"
)
head(atlantic_polio_1950s)
#> # A tibble: 6 × 21
#>   iso_3166 iso_3166_2 period_start_date period_end_date disease  nesting_disease
#>   <chr>    <chr>      <date>            <date>          <chr>    <chr>          
#> 1 CA       CA-NL      1950-01-01        1950-01-07      poliomy… poliomyelitis  
#> 2 CA       CA-NL      1950-01-01        1950-01-07      poliomy… poliomyelitis  
#> 3 CA       CA-NL      1950-01-01        1950-01-07      poliomy… poliomyelitis  
#> 4 CA       CA-NL      1950-01-08        1950-01-14      poliomy… poliomyelitis  
#> 5 CA       CA-NL      1950-01-08        1950-01-14      poliomy… poliomyelitis  
#> 6 CA       CA-NL      1950-01-08        1950-01-14      poliomy… poliomyelitis  
#> # ℹ 15 more variables: basal_disease <chr>, cases_this_period <dbl>,
#> #   location <chr>, location_type <chr>, time_scale <chr>,
#> #   historical_disease <chr>, historical_disease_family <chr>,
#> #   historical_disease_subclass <chr>, days_this_period <dbl>,
#> #   period_mid_date <date>, population <dbl>, record_origin <chr>,
#> #   original_dataset_id <chr>, digitization_id <chr>, scan_id <chr>