iidda.api

IIDDA API.

Details

IIDDA is the International Infectious Disease Data Archive. This archive has an API (Application Programming Interface) for accessing data and potentially for building applications. This R package provides a simple wrapper to this API so that the datasets are returned as data frames.

Listing the Datasets

Getting Datasets

Just choose one of the identifiers above, and pass it to the raw_csv function in the API.

You can pass more than one dataset.

All Metadata by Dataset

On my browser at least, it is easier to look at in JSON form rather than R list form. The JSON metadata can be directly accessed using the URL form of the API: https://math.mcmaster.ca/iidda/api/metadata.

CANMOD Digitization Project

The communicable disease incidence (CDI) data collected as part of the CANMOD digitization project can be accessed using resource_type = "CANMOD CDI" with the filter function.

Author

Maintainer: Steve Walker swalk@mcmaster.ca