IIDDA API Operations

Objects containing the functions associated with API functions documented here. These objects are for advanced usage, providing more functionality than the simpler tools for accessing featured_data.

Usage

ops

ops_local

ops_staging

Format

An object of class try-error of length 1.

An object of class try-error of length 1.

An object of class list of length 6.

Functions

ops: List containing available operations from the IIDDA API as R functions

ops_local: Operations list for a local development environment, if it exists

ops_staging: Operations list for a staging environment, if it exists

Examples

## Print out the available functions.
names(ops_staging)
#> [1] "metadata"        "data_dictionary" "lookup_tables"   "raw_csv"        
#> [5] "download"        "filter"         

## Access functions with a dollar sign. For example, this command
## will give unharmonized weekly incidence data in January of 1956.
options(iidda_api_msgs = FALSE)
ops_staging$filter(
     resource_type = "Communicable Disease Incidence"
   , dataset_id = "cdi_ca_1956_wk_prov_dbs"
   , period_end_date = "1956-01-01..1956-02-01"
   , time_scale = "wk"
)
#> Warning: 66 parsing failures.
#> row col expected         actual
#>  27  -- a number Not reportable
#>  38  -- a number Not reportable
#>  39  -- a number Not reportable
#>  49  -- a number Not reportable
#>  50  -- a number Not reportable
#> ... ... ........ ..............
#> See problems(...) for more details.
#> Warning: 76 parsing failures.
#> row col expected         actual
#>  10  -- a number Not reportable
#>  27  -- a number Not reportable
#>  38  -- a number Not reportable
#>  39  -- a number Not reportable
#>  49  -- a number Not reportable
#> ... ... ........ ..............
#> See problems(...) for more details.
#> Warning: 74 parsing failures.
#> row col expected         actual
#>   9  -- a number Not reportable
#>  10  -- a number Not reportable
#>  27  -- a number Not reportable
#>  38  -- a number Not reportable
#>  39  -- a number Not reportable
#> ... ... ........ ..............
#> See problems(...) for more details.
#> # A tibble: 1,232 × 16
#>    cases_cum_median_prev_5_years cases_cum_prev_year cases_cum_report_year
#>                            <dbl>               <dbl>                 <dbl>
#>  1                          1415                 921                   927
#>  2                            NA                  NA                    NA
#>  3                            NA                  NA                    NA
#>  4                            NA                  NA                    NA
#>  5                            NA                  NA                    NA
#>  6                            NA                  NA                    NA
#>  7                            NA                  NA                    NA
#>  8                            NA                  NA                    NA
#>  9                            NA                  NA                    NA
#> 10                            NA                  NA                    NA
#> # ℹ 1,222 more rows
#> # ℹ 13 more variables: cases_median_prev_5_years <dbl>,
#> #   cases_prev_period <dbl>, cases_this_period <dbl>, dataset_id <chr>,
#> #   digitization_id <chr>, historical_disease <chr>,
#> #   historical_disease_subclass <chr>, location <chr>, location_type <chr>,
#> #   period_end_date <date>, period_start_date <date>, scan_id <chr>,
#> #   time_scale <chr>

## Operations objects that are not available are error objects. As of the
## time of writing `ops` is not not live, but will be.
print(class(ops))
#> [1] "try-error"

## The `ops_local` is only live for developers who have deployed a
## local version of the API.
print(class(ops_local))
#> [1] "try-error"