iidda.api

IIDDA API.

Details

IIDDA is the International Infectious Disease Data Archive. This archive has an API (Application Programming Interface) for accessing data and potentially for building applications. This R package provides a simple wrapper to this API so that the datasets are returned as data frames.

Listing the Datasets

iidda.api::ops_staging$metadata() |> names()
#>   [1] "cdi_on_1990-2021_wk"
#>   [2] "cdi_ca_1978_wk_prov_statcan"
#>   [3] "cdi_ca_1977_wk_prov_statcan"
#>   [4] "cdi_ca_1976_wk_prov_statcan"
#>   [5] "cdi_ca_1975_wk_prov_statcan"
#>   [6] "cdi_wc_ca_1924-55_wk_prov"
#>   [7] "cdi_ca_1993-2001_quart_prov_hc"
#>   [8] "cdi_ca_1990-1992_mn_prov_hc"
#>   [9] "cdi_ca_1984-89_4wk_prov_statcan"
#>  [10] "cdi_ca_1979-83_4wk_prov_statcan"
#>  [11] "cdi_ca_1972_wk_prov"
#>  [12] "cdi_ca_1971_wk_prov"
#>  [13] "cdi_ca_1970_wk_prov"
#>  [14] "cdi_ca_1969_wk_prov"
#>  [15] "cdi_ca_1968_wk_prov"
#>  [16] "cdi_ca_1967_wk_prov_mcgill"
#>  [17] "cdi_ca_1966_wk_prov_mcgill"
#>  [18] "cdi_ca_1965_wk_prov_mcgill"
#>  [19] "cdi_ca_1964_wk_prov_mcgill"
#>  [20] "cdi_wc_ca_1924_wk_prov"
#>  [21] "mort_on_1940-1947_mn"
#>  [22] "mort_on_1903-1939_mn"
#>  [23] "cdi_sk_1910-1927_mn"
#>  [24] "cdi_on_1940-1947_mn"
#>  [25] "cdi_on_1903-1939_mn"
#>  [26] "cdi_ca_1962_wk_prov_dbs"
#>  [27] "cdi_ab_2004-19_wk_age"
#>  [28] "cdi_ca_1903-2021_wk-qr_prov_normalized"
#>  [29] "cdi_yf_ca_1945-1955_wk_prov"
#>  [30] "cdi_qc_1915-1925_mn_county_municipality"
#>  [31] "cdi_on_1939-1989_wk"
#>  [32] "cdi_granuing_ca_1946-55_wk_prov"
#>  [33] "cdi_ca_1973_wk_prov"
#>  [34] "cdi_ca_1960_wk_prov_dbs"
#>  [35] "pop_ca_1921-71_an_age_prov_sex"
#>  [36] "mort_ca_2000-2020_wk_prov_causes"
#>  [37] "mort_ca_1990-1999_wk_prov_causes"
#>  [38] "mort_ca_1980-1989_wk_prov_causes"
#>  [39] "mort_ca_1970-1979_wk_prov_causes"
#>  [40] "mort_ca_1960-1969_wk_prov_causes"
#>  [41] "mort_ca_1950-1959_wk_prov_causes"
#>  [42] "cdi_ca_1924-55_wk_prov_dbs_statcan_harmonized"
#>  [43] "cdi_on_1990-2021_wk_harmonized"
#>  [44] "cdi_ca_1979-89_4wk_prov_harmonized"
#>  [45] "cdi_tball_ca_1924-55_wk_prov"
#>  [46] "cdi_chanc_ca_1924-55_wk_prov"
#>  [47] "cdi_ca_1990-2001_quart_prov_harmonized"
#>  [48] "cdi_ca_1975-78_wk_prov_harmonized"
#>  [49] "cdi_ca_1961_wk_prov_dbs"
#>  [50] "cdi_ca_1959_wk_prov_dbs_w4_52"
#>  [51] "cdi_ca_1959_wk_prov_dbs_w1_3"
#>  [52] "cdi_ca_1958_wk_prov_dbs"
#>  [53] "cdi_ca_1957_wk_prov_dbs"
#>  [54] "cdi_ca_1956_wk_prov_dbs"
#>  [55] "cdi_bot_ca_1933-55_wk_prov"
#>  [56] "cdi_ab_2004-19_wk_harmonized"
#>  [57] "cdi_uf_ca_1928-55_wk_prov"
#>  [58] "cdi_typhpara_ca_1952-55_wk_prov"
#>  [59] "cdi_typh_ca_1924-52_wk_prov"
#>  [60] "cdi_tbunspec_ca_1933-55_wk_prov"
#>  [61] "cdi_tbpulm_ca_1933-55_wk_prov"
#>  [62] "cdi_tbmisc_ca_1933-55_wk_prov"
#>  [63] "cdi_syphunspec_ca_1944-55_wk_prov"
#>  [64] "cdi_syphtertother_ca_1944-55_wk_prov"
#>  [65] "cdi_syphtertneuro_ca_1944-55_wk_prov"
#>  [66] "cdi_syphtertcard_ca_1944-55_wk_prov"
#>  [67] "cdi_syphsec_ca_1944-55_wk_prov"
#>  [68] "cdi_syphprim_ca_1944-55_wk_prov"
#>  [69] "cdi_syphother_ca_1944-51_wk_prov"
#>  [70] "cdi_syphlatent_ca_1944-55_wk_prov"
#>  [71] "cdi_syphcong_ca_1944-55_wk_prov"
#>  [72] "cdi_syphall_ca_1924-55_wk_prov"
#>  [73] "cdi_sst_ca_1924-55_wk_prov"
#>  [74] "cdi_smallpox_ca_1924-55_wk_prov"
#>  [75] "cdi_sf_ca_1924-55_wk_prov"
#>  [76] "cdi_salm_ca_1944-55_wk_prov"
#>  [77] "cdi_rub_ca_1924-55_wk_prov"
#>  [78] "cdi_poliopar_ca_1949-55_wk_prov"
#>  [79] "cdi_polio_ca_1924-55_wk_prov"
#>  [80] "cdi_pneu_ca_1924-55_wk_prov"
#>  [81] "cdi_paratyph_ca_1924-55_wk_prov"
#>  [82] "cdi_optneo_ca_1944-55_wk_prov"
#>  [83] "cdi_on_1939-1989_wk_harmonized"
#>  [84] "cdi_nspecurethritis_ca_1949-55_wk_prov"
#>  [85] "cdi_mumps_ca_1924-55_wk_prov"
#>  [86] "cdi_meningosep_ca_1952-54_wk_prov"
#>  [87] "cdi_meningivar_ca_1925-55_wk_prov"
#>  [88] "cdi_meningisep_ca_1935-55_wk_prov"
#>  [89] "cdi_malar_ca_1929-55_wk_prov"
#>  [90] "cdi_lymphgran_ca_1946-55_wk_prov"
#>  [91] "cdi_jauncat_ca_1948-55_wk_prov"
#>  [92] "cdi_jaun_ca_1927-55_wk_prov"
#>  [93] "cdi_impet_ca_1924-55_wk_prov"
#>  [94] "cdi_hep_ca_1945-55_wk_prov"
#>  [95] "cdi_erys_ca_1924-55_wk_prov"
#>  [96] "cdi_enceph_ca_1924-34_wk_prov"
#>  [97] "cdi_dysunspec_ca_1927-55_wk_prov"
#>  [98] "cdi_dysbac_ca_1924-55_wk_prov"
#>  [99] "cdi_dysam_ca_1927-55_wk_prov"
#> [100] "cdi_dysall_ca_1929-55_wk_prov"
#> [101] "cdi_diphth_ca_1924-55_wk_prov"
#> [102] "cdi_diarent_ca_1945-55_wk_prov"
#> [103] "cdi_diar_ca_1937-44_wk_prov"
#> [104] "cdi_cp_ca_1924-55_wk_prov"
#> [105] "cdi_ca_1968-72_wk_prov_harmonized"
#> [106] "cdi_ca_1964-67_wk_prov_harmonized"
#> [107] "cdi_tbnpulm_ca_1935-55_wk_prov"
#> [108] "cdi_qc_1915-1925_mn_harmonized"
#> [109] "cdi_poliounspec_ca_1949-55_wk_prov"
#> [110] "cdi_polionpar_ca_1949-55_wk_prov"
#> [111] "cdi_otherVD_ca_1944-55_wk_prov"
#> [112] "cdi_on_1940-1947_mn_harmonized"
#> [113] "cdi_on_1903-1939_mn_harmonized"
#> [114] "cdi_gonall_ca_1924-55_wk_prov"
#> [115] "cdi_ca_1956-63_1973-74_wk_prov_harmonized"
#> [116] "prototype-dataset"
#> [117] "pop_ca_1971-2021_an_age_prov_sex"
#> [118] "pop_ca_1871-2021_harmonized"
#> [119] "pop_ca_1871-1921_10-yearly_prov_age_sex"
#> [120] "cdi_sk_1910-1927_mn_harmonized"
#> [121] "cdi_ca_1963_wk_prov_dbs"

Getting Datasets

Just choose one of the identifiers above, and pass it to the raw_csv function in the API.

cdi_1975 = iidda.api::ops_staging$raw_csv(
   dataset_ids = c(
     "cdi_ca_1975_wk_prov_statcan",
     "cdi_ca_1976_wk_prov_statcan"
   )
)

You can pass more than one dataset.

cdi_1975 = iidda.api::ops_staging$raw_csv(
   dataset_ids = "cdi_ca_1975_wk_prov_statcan"
)

All Metadata by Dataset

iidda.api::ops_staging$metadata()

On my browser at least, it is easier to look at in JSON form rather than R list form. The JSON metadata can be directly accessed using the URL form of the API: https://math.mcmaster.ca/iidda/api/metadata.

CANMOD Digitization Project

The communicable disease incidence (CDI) data collected as part of the CANMOD digitization project can be accessed using resource_type = "CANMOD CDI" with the filter function.

canmod_cdi = iidda.api::ops_staging$filter(
     resource_type = "CANMOD CDI"
   , iso_3166 = "CA" ## country code for canada
)

Author

Maintainer: Steve Walker swalk@mcmaster.ca