IIDDA API Operations

Objects containing the functions associated with API functions documented here.

Usage

ops

ops_local

ops_staging

Format

An object of class try-error of length 1.

An object of class try-error of length 1.

An object of class list of length 6.

Functions

ops: List containing available operations from the IIDDA API as R functions

ops_local: Operations list for a local development environment, if it exists

ops_staging: Operations list for a staging environment, if it exists

Examples

## Print out the available functions.
names(ops_staging)
#> [1] "metadata"        "data_dictionary" "lookup_tables"   "raw_csv"        
#> [5] "download"        "filter"         

## Access functions with a dollar sign. For example, this command
## will give weekly incidence data in PEI in January of 1940.
ops_staging$filter(
     resource_type = "CANMOD CDI"
   , iso_3166_2 = "CA-PE"
   , period_end_date = "1940-01-01..1940-02-01"
   , time_scale = "wk"
)
#> Now converting some fields from character to numeric or date. 
#> You can turn this off with
#> options(iidda_api_all_char = TRUE). 
#> Do not display any iidda.api messages with
#> options(iidda_api_msgs = FALSE).
#> Messages displayed at most once per R session.
#> Now sorting by date. 
#> You can turn this off with
#> options(iidda_api_date_sort = FALSE). 
#> Do not display any iidda.api messages with
#> options(iidda_api_msgs = FALSE).
#> Messages displayed at most once per R session.
#> # A tibble: 68 × 20
#>    iso_3166 iso_3166_2 location location_type period_start_date period_end_date
#>    <chr>    <chr>      <chr>    <chr>         <date>            <date>         
#>  1 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  2 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  3 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  4 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  5 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  6 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  7 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  8 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#>  9 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#> 10 CA       CA-PE      P.E.I    province      1939-12-31        1940-01-06     
#> # ℹ 58 more rows
#> # ℹ 14 more variables: period_mid_date <date>, days_this_period <dbl>,
#> #   time_scale <chr>, disease <chr>, nesting_disease <chr>,
#> #   basal_disease <chr>, historical_disease_family <chr>,
#> #   historical_disease <chr>, historical_disease_subclass <chr>,
#> #   dataset_id <chr>, original_dataset_id <chr>, cases_this_period <dbl>,
#> #   population <dbl>, population_reporting <dbl>

## Operations objects that are not available are error objects. As of the
## time of writing only ops_staging is live.
print(ops)
#> [1] "Error in do.call(make_ops_list, production) : \n  second argument must be a list\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in do.call(make_ops_list, production): second argument must be a list>
print(ops_local)
#> [1] "Error in rapiclient::get_api(url = file.path(iidda::rm_trailing_slash(file.path(api_url,  : \n  'url' does not appear to be JSON or YAML\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in rapiclient::get_api(url = file.path(iidda::rm_trailing_slash(file.path(api_url,     base_path)), "openapi.json")): 'url' does not appear to be JSON or YAML>