IIDDA API Operations

Objects containing the functions associated with API functions documented here.

Usage

ops

ops_local

ops_staging

Format

An object of class try-error of length 1.

An object of class try-error of length 1.

An object of class list of length 6.

Functions

ops: List containing available operations from the IIDDA API as R functions

ops_local: Operations list for a local development environment, if it exists

ops_staging: Operations list for a staging environment, if it exists

Examples

## Print out the available functions.
names(ops_staging)
#> [1] "metadata"        "data_dictionary" "lookup_tables"   "raw_csv"        
#> [5] "download"        "filter"         

## Access functions with a dollar sign. For example, this command
## will give weekly incidence data in PEI in January of 1940.
ops_staging$filter(
     resource_type = "CANMOD CDI"
   , iso_3166_2 = "CA-PE"
   , period_end_date = "1940-01-01..1940-02-01"
   , time_scale = "wk"
)
#> Now converting some fields from character to numeric or date. 
#> You can turn this off with
#> options(iidda_api_all_char = TRUE). 
#> Do not display any iidda.api messages with
#> options(iidda_api_msgs = FALSE).
#> Messages displayed at most once per R session.
#> Now sorting by date. 
#> You can turn this off with
#> options(iidda_api_date_sort = FALSE). 
#> Do not display any iidda.api messages with
#> options(iidda_api_msgs = FALSE).
#> Messages displayed at most once per R session.
#> # A tibble: 68 × 16
#>    basal_disease         cases_this_period dataset_id disease historical_disease
#>    <chr>                             <dbl> <chr>      <chr>   <chr>             
#>  1 botulism                              0 cdi_ca_19… botuli… Botulism          
#>  2 chickenpox                            0 cdi_ca_19… chicke… Chickenpox        
#>  3 diphtheria                            0 cdi_ca_19… diphth… Diphtheria        
#>  4 malaria                               0 cdi_ca_19… malaria Malaria           
#>  5 mumps                                 0 cdi_ca_19… mumps   Mumps             
#>  6 typhoid-paratyphoid-…                 0 cdi_ca_19… paraty… Paratyphoid Fever 
#>  7 pneumonia                             3 cdi_ca_19… pneumo… Pneumonia         
#>  8 poliomyelitis                         0 cdi_ca_19… poliom… Poliomyelitis     
#>  9 rubella                               0 cdi_ca_19… rubella Rubella(German me…
#> 10 scarlet-fever                         8 cdi_ca_19… scarle… Scarlet Fever     
#> # ℹ 58 more rows
#> # ℹ 11 more variables: historical_disease_family <chr>,
#> #   historical_disease_subclass <chr>, iso_3166 <chr>, iso_3166_2 <chr>,
#> #   location <chr>, location_type <chr>, nesting_disease <chr>,
#> #   original_dataset_id <chr>, period_end_date <date>,
#> #   period_start_date <date>, time_scale <chr>

## Operations objects that are not available are error objects. As of the
## time of writing only ops_staging is live.
print(ops)
#> [1] "Error in do.call(make_ops_list, production) : \n  second argument must be a list\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in do.call(make_ops_list, production): second argument must be a list>
print(ops_local)
#> [1] "Error in rapiclient::get_api(url = file.path(iidda::rm_trailing_slash(file.path(api_url,  : \n  'url' does not appear to be JSON or YAML\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in rapiclient::get_api(url = file.path(iidda::rm_trailing_slash(file.path(api_url,     base_path)), "openapi.json")): 'url' does not appear to be JSON or YAML>